backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
python-docx>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0
python-dateutil>=2.8.0
mammoth>=1.6.0
openpyxl>=3.1.0  # only for tests (creating Excel files)
python-calamine>=0.1.0  # fast Excel reader (Rust-based)
pytest>=7.4.0
pytest-asyncio>=0.21.0
